// src/app/model/candidate.model.ts
export class Candidate {
  id? : number;
  CIN? : number;
  firstname?: String;
  lastname?: String;
  bacgeneration?: number;
  moyen?: number;
  address?: String;
  cityOfOrigin?: String;
  zipOfOrigin?: String;
  stateOfOrigin?: String;
  cityOfStudy?: String;
  stateOfStudy?: String;
  zipOfStudy?: String;
  sexe?: String;
  status?: String;
}